project/requirements.txt
requests
python-frontmatter
openai
python-dotenv
minsearch 
pydantic-ai
streamlit